# Conflicts:
#	case_module6/src/main/java/com/example/case_module6/controller/DrinkRestController.java
#	case_module6/src/main/java/com/example/case_module6/controller/InvoiceRestController.java
#	case_module6/src/main/java/com/example/case_module6/repository/CartRepository.java
#	case_module6/src/main/java/com/example/case_module6/repository/OrderDetailRepository.java
#	case_module6/src/main/java/com/example/case_module6/service/IInvoiceService.java
#	case_module6/src/main/java/com/example/case_module6/service/implement/DrinkService.java
#	case_module6/src/main/java/com/example/case_module6/service/implement/InvoiceService.java

+++ case_module6/src/main/java/com/example/case_module6/repository/CartRepository.java
@@ -2,6 +2,5 @@ package com.example.case_module6.repository;
 
 import com.example.case_module6.model.Cart;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CartRepository extends JpaRepository<Cart, Integer> {
+public interface CartRepository extends JpaRepository<Cart, Long> {
 }
